+++ R/give.R
@@ -1,6 +1,6 @@
 #' Extract information from \code{powRICLPM} object
 #'
-#' \code{give()} can be used to extract information stored within a \code{powRICLPM} object. See "Details" for which pieces of information can be extracted. The information is presented by condition (i.e., sample size, number of time points, and ICC).
+#' \code{give()} can be used to extract information stored within a \code{powRICLPM} object, and is internally used by \code{\link{print.powRICLPM}} and \code{\link{summary.powRICLPM}}. See "Details" for which pieces of information can be extracted. The information is presented by condition (i.e., sample size, number of time points, and ICC).
 #'
 #' @param from A \code{powRICLPM} object
 #' @param what A character string denoting the information to extract, either "conditions", "estimation_problems", "results", or "names".
@@ -12,7 +12,7 @@
 #' \itemize{
 #'   \item \code{conditions}: A \code{data.frame} with the different experimental conditions per row, where each condition is defined by a unique combination of sample size, number of time points and ICC.
 #'   \item \code{estimation_problems}: The proportion of fatal errors, inadmissible values, or non-converged estimations (columns) per experimental conditions (row).
-#'   \item \code{results}: The average estimate (\code{Avg}), minimum estimate (\code{Min}), standard deviation of parameter estimates (\code{stdDev}), the average standard error (\code{SEavg}), the mean square error (\code{MSE}), the average width of the confidence interval (\code{Acc}), the coverage rate (\code{Cov}), and the proportion of times the *p*-value was lower than the significance criterion (\code{Pow}). It requires setting the \code{parameter = "..."} argument.
+#'   \item \code{results}: The average estimate (\code{Avg}), minimum estimate (\code{Min}), standard deviation of parameter estimates (\code{stdDev}), the average standard error (\code{SEavg}), the mean square error (\code{MSE}), the average width of the confidence interval (\code{Acc}), the coverage rate (\code{Cov}), and the proportion of times the \emph{p}-value was lower than the significance criterion (\code{Pow}). It requires setting the \code{parameter = "..."} argument.
 #'   \item \code{names}: The parameter names in the condition with the least parameters (i.e., parameter names that apply to each experimental condition).
 #' }
 #'