+++ R/powRICLPM.R
@@ -1,41 +1,46 @@
 #' @title
-#' Power analysis for the RI-CLPM
+#' Power analysis for the RI-CLPM (and STARTS model)
 #'
 #' @description
-#' \code{powRICLPM()} performs a Monte Carlo power analysis for the random intercept cross-lagged panel model (RI-CLPM). It computes performance metrics (e.g., bias, mean square error, coverage, power, etc) for all RI-CLPM parameters, and can perform power analysis across multiple experimental conditions simultaneously. Conditions are defined in terms of sample size, number of time points, and proportion of between-unit variance (ICC). See "Details" for information on a) the data simulation and model estimation, b) internal naming conventions of parameters, c) the option to include measurement errors and various constraints, and d) parallel execution capabilities for speeding up the analysis.
+#' \code{powRICLPM()} performs a Monte Carlo power analysis for the random intercept cross-lagged panel model (RI-CLPM). It computes performance metrics (e.g., bias, mean square error, coverage, power, etc) for all RI-CLPM parameters, and can perform power analyses across multiple experimental conditions simultaneously. Conditions are defined in terms of sample size, number of time points, and proportion of between-unit variance (ICC). See "Details" for information on a) the data simulation, b) model estimation, c) internal naming conventions of parameters, d) the option to include measurement errors (i.e., estimating the Stable Trait Autoregressive Trait State model), e) imposing various constraints over time, and f) parallel execution capabilities for speeding up the analysis.
 #'
 #' @param target_power A numeric value between 0 and 1, denoting the targeted power level.
 #' @param search_lower A positive integer, denoting the lower bound of a range of sample sizes.
 #' @param search_upper A positive integer, denoting the upper bound of a range of sample sizes.
 #' @param search_step A positive integer, denoting an increment in sample size.
-#' @param sample_size (optional) An integer (vector), indicating specific sample sizes at which to evaluate power, rather than specifying a range using the \code{search_} arguments.
+#' @param sample_size (optional) An integer (vector), indicating specific sample sizes at which to evaluate power, rather than specifying a range using the \code{search_*} arguments.
 #' @param time_points An integer (vector) with elements at least larger than 3, indicating number of time points.
-#' @param ICC A \code{double} (vector), denoting the proportion of variance at the between-unit level.
-#' @param RI_cor A \code{double}, denoting the correlation between random intercepts.
+#' @param ICC A \code{double} (vector) with elements between 0 and 1, denoting the proportion of (true score) variance at the between-unit level. When measurement error is included in the data generating model, ICC is computed as the variance of the random intercept factor divided by the true score variance (i.e., controlled for measurement error).
+#' @param RI_cor A \code{double} between 0 and 1, denoting the correlation between random intercepts.
 #' @param Phi A matrix, with standardized autoregressive effects (on the diagonal) and cross-lagged effects (off-diagonal) in the population. Columns represent predictors and rows represent outcomes.
-#' @param wSigma A correlation matrix for the within-unit components.
+#' @param within_cor A \code{double} between 0 and 1, denoting the correlation between the within-unit components.
 #' @param reliability (optional) A numeric value between 0 and 1, denoting the reliability of the variables.
 #' @param skewness (optional) A numeric value, denoting the skewness values for the observed variables (see \code{\link[lavaan]{simulateData}}).
 #' @param kurtosis (optional) A numeric value, denoting the excess kurtosis values (i.e., compared to the kurtosis of a normal distribution) for the observed variables (see \code{\link[lavaan]{simulateData}}).
-#' @param est_ME (optional) A logical, denoting if measurement error variance should be estimated in the RI-CLPM (see "Details").
+#' @param estimate_ME (optional) A logical, denoting if measurement error variance should be estimated in the RI-CLPM (see "Details").
 #' @param alpha (optional) A \code{double}, denoting the significance criterion.
 #' @param reps A positive \code{integer}, denoting the number of Monte Carlo replications to be used during simulations.
-#' @param bootstrap_reps (optional) A positive \code{integer}, denoting the number of bootstrap samples to use for quantifying the uncertainty around the power analysis results.
-#' @param seed An `integer` of length 1. If multiple cores are used, a seed of length 1 will be used to generate a full L'Ecuyer-CMRG seed for all cores (see \code{\link[furrr]{furrr_options}}).
+#' @param bootstrap_reps (optional) A positive \code{integer}, denoting the number of bootstrap samples to use for quantifying the uncertainty (i.e., 95\% bootstrap confidence interval) around the power analysis results.
+#' @param seed An \code{integer} of length 1. If multiple cores are used, a seed of length 1 will be used to generate a full L'Ecuyer-CMRG seed for all cores (see \code{\link[furrr]{furrr_options}}).
 #' @param constraints (optional) A character string, specifying the type of constraints that should be imposed on the estimation model (see "Details").
 #' @param bounds (optional) A logical, denoting if bounded estimation should be used for the latent variable variances in the model (see "Details").
-#' @param estimator (options) A character, denoting the estimator to be used (see "Details").
+#' @param estimator (options) A character, denoting the estimator to be used (default: \code{ML}, see "Details").
+#' @param save_path A character string naming the directory to save (data) files to (used for validation purposes of this package). Variables are saved in alphabetical and numerical order.
 #'
 #' @details
 #' A rationale for the power analysis strategy implemented in this package can be found in Mulder (2022).
 #'
-#' \subsection{Data generation and model estimation}{Data are generated using \code{\link[lavaan]{simulateData}} from the \pkg{lavaan} package. Based on \code{Phi} and \code{wSigma}, the residual variances and covariances for the within-components at wave 2 and later are computed, such that the within-components themselves have a variance of 1. This implies that the lagged effects in \code{Phi} can be interpreted as standardized effects.
+#' \subsection{Data generation}{Data are generated using \code{\link[lavaan]{simulateData}} from the \pkg{lavaan} package. Based on \code{Phi} and \code{within_cor}, the residual variances and covariances for the within-components at wave 2 and later are computed, such that the within-components themselves have a variance of 1. This implies that the lagged effects in \code{Phi} can be interpreted as standardized effects.}
 #'
-#' Data are analyzed using \code{\link[lavaan]{lavaan}} from the \pkg{lavaan} package. The default estimator is maximum likelihood. Other maximum likelihood based estimators implemented in \href{https://lavaan.ugent.be/tutorial/est.html}{\pkg{lavaan}} can be specified as well. When skewed or kurtosed data are generated (using the \code{skewness} and \code{kurtosis} arguments), the estimator defaults to robust maximum likelihood (i.e., \code{estimator = "MLR"}). The population parameter values are used as starting values. Parameter estimates from unconverged model solutions are discarded from the results. However, parameter estimates from solutions with inadmissible parameter estimates (e.g., a negative random intercept variance), are included in the results. The results include the minimum estimates for all parameters across replications. This can be used to diagnose which parameter is the cause of the inadmissible solution.}
+#' \subsection{Model estimation}{Data are analyzed using \code{\link[lavaan]{lavaan}} from the \pkg{lavaan} package. The default estimator is maximum likelihood (\code{ML}). Other maximum likelihood based estimators implemented in \href{https://lavaan.ugent.be/tutorial/est.html}{\pkg{lavaan}} can be specified as well. When skewed or kurtosed data are generated (using the \code{skewness} and \code{kurtosis} arguments), the estimator defaults to robust maximum likelihood \code{MLR}. The population parameter values are used as starting values.
+#'
+#' Parameter estimates from non-converged model solutions are discarded from the results. When \code{bounds = FALSE}, inadmissible parameter estimates from converged solutions (e.g., a negative random intercept variance) are discarded. When \code{bounds = TRUE}, inadmissible parameter estimates are retained following advice by \href{https://doi.org/10.1080/10705511.2021.1982716}{De Jonckere and Rosseel (2022)}. The results include the minimum estimates for all parameters across replications to diagnose which parameter(s) is (are) the cause of the inadmissible solution.}
 #'
 #' \subsection{Naming conventions for observed and latent variables}{The observed variables in the RI-CLPM are given default names, namely capital letters in alphabetical order, with numbers denoting the measurement occasion. For example, for a bivariate RICLPM with 3 time points, we observe \code{A1}, \code{A2}, \code{A3}, \code{B1}, \code{B2}, and \code{B3}. Their within-components are denoted by \code{wA1}, \code{wA2}, ..., \code{wB3}, respectively. The between-components have \code{RI_} prepended to the variable name, resulting in \code{RI_A} and \code{RI_B}.
 #'
 #' Parameters are denoted using \pkg{lavaan} model syntax (see \href{https://lavaan.ugent.be/tutorial/syntax1.html}{the \pkg{lavaan} website}). For example, the random intercept variances are denoted by \code{RI_A~~RI_A} and \code{RI_B~~RI_B}, the cross-lagged effects at the first wave as \code{wB2~wA1} and \code{wA2~wB1}, and the autoregressive effects as \code{wA2~wA1} and \code{wB2~wB1}. Use \code{give(object, "names")} to extract parameter names from the \code{powRICLPM} object.}
+#'
+#' \subsection{Measurement errors (STARTS model)}{Including measurement error to the RI-CLPM makes the model equivalent to the Stable Trait Autoregressive Trait State (STARTS) model by \href{https://doi.org/10.1037/10409-008}{Kenny and Zautra (2001)} without constraints over time. Measurement error can be added to the generated data through the \code{reliability} argument. Setting \code{reliability = 0.8} implies that 80\% is true score variance and 20\% is measurement error variance; \code{ICC} then denotes the proportion of \emph{true score variance} captured by the random intercept factors. Estimating measurement errors (i.e., the STARTS model) is done by setting \code{est_ME = TRUE}.}
 #'
 #' \subsection{Imposing constraints}{The following constraints can be imposed on the estimation model using the \code{constraints = "..."} argument:
 #'
@@ -44,15 +49,15 @@
 #'   \item \code{residuals}: Time-invariant residual variances.
 #'   \item \code{within}: Time-invariant lagged effects and residual variances.
 #'   \item \code{stationarity}: Constraints such that at the within-unit level a stationary process is estimated. This included time-invariant lagged effects, and constraints on the residual variances.
-#'   \item \code{ME}: Time-invariant measurement error variances. Only possible when \code{est_ME = TRUE}.
+#'   \item \code{ME}: Time-invariant measurement error variances. Only possible when \code{estimate_ME = TRUE}.
 #' }
 #' }
 #'
-#' \subsection{Bounded estimation}{Bounded estimation is useful to avoid nonconvergence in small samples (< 100). Here, automatic wide bounds are used as advised by \doi{https://doi.org/10.1080/10705511.2021.1982716}{De Jonckere and Rosseel (2022}; see \code{optim.bounds} in \code{\link[lavaan]{lavOptions}}). This option can only be used when no constraints are imposed on the estimation model.}
+#' \subsection{Bounded estimation}{Bounded estimation is useful to avoid nonconvergence in small samples (< 100). Here, automatic wide bounds are used as advised by \href{https://doi.org/10.1080/10705511.2021.1982716}{De Jonckere and Rosseel (2022)}; See \code{optim.bounds} in \code{\link[lavaan]{lavOptions}}. This option can only be used when no constraints are imposed on the estimation model.}
 #'
-#' \subsection{Parallel processing using \pkg{furrr}}{To speed up the analysis, power analysis for multiple experimental conditions can be executed in parallel. This has been implemented using \pkg{furrr}. By default the analysis is executed sequentially (i.e., single-core). Parallel execution (i.e., multicore) can be setup using \code{\link[future]{plan}}, for example \code{plan(multisession, workers = 4)}. For more information and options, see \url{https://furrr.futureverse.org}.}
+#' \subsection{Parallel processing using \pkg{furrr}}{To speed up the analysis, power analysis for multiple experimental conditions can be executed in parallel. This has been implemented using \pkg{furrr}. By default the analysis is executed sequentially (i.e., single-core). Parallel execution (i.e., multicore) can be setup using \code{\link[future]{plan}}, for example \code{plan(multisession, workers = 4)}. For more information and options, see \url{https://furrr.futureverse.org}.
 #'
-#' \subsection{Progress bar using \pkg{progressr}}{A progress bar displaying the status of the power analysis has been implemented using \pkg{progressr}. By default, a simple progress bar will be shown. For more information on how to control this progress bar and several other notification options (e.g., auditory notifications), see \url{https://progressr.futureverse.org}.}
+#' A progress bar displaying the status of the power analysis has been implemented using \pkg{progressr}. By default, a simple progress bar will be shown. For more information on how to control this progress bar and several other notification options (e.g., auditory notifications), see \url{https://progressr.futureverse.org}.}
 #'
 #' @return
 #' A list containing a \code{conditions} and \code{session} element. \code{condition} itself is a list of experimental conditions, where each element is again a list containing the input and output of the power analysis for that particular experimental condition. \code{session} is a list containing information common to all experimental conditions.
@@ -70,9 +75,8 @@
 #'
 #' @examples
 #' # Example - Simulate power across range of sample sizes
-#' # Define population parameters for lagged effects and within-component correlations
+#' # Define population parameters for lagged effects
 #' Phi <- matrix(c(.4, .1, .2, .3), ncol = 2, byrow = TRUE)
-#' wSigma <- matrix(c(1, .3, .3, 1), ncol = 2, byrow = TRUE)
 #'
 #' # Setup parallel computing (multicore, speeding up the analysis)
 #' \dontrun{
@@ -91,7 +95,7 @@
 #'     ICC = c(0.4, 0.5, 0.6),
 #'     RI_cor = 0.3,
 #'     Phi = Phi,
-#'     wSigma = wSigma,
+#'     within_cor = 0.3,
 #'     reps = 50,
 #'     seed = 1234
 #'   )
@@ -106,40 +110,42 @@ powRICLPM <- function(target_power,
                       ICC,
                       RI_cor,
                       Phi,
-                      wSigma,
+                      within_cor,
                       reliability = 1,
                       skewness = 0,
                       kurtosis = 0,
-                      est_ME = FALSE,
+                      estimate_ME = FALSE,
                       alpha = 0.05,
                       reps = 20,
                       bootstrap_reps = 1000,
                       seed = NA,
                       constraints = "none",
                       bounds = FALSE,
-                      estimator = NA) {
+                      estimator = NA,
+                      save_path = NULL) {
 
   # Check arguments I
   message(rlang::format_error_bullets(c(
     i = "Checking arguments..."
   )))
   target_power <- check_target(target_power)
-  time_points <- check_T(time_points, est_ME)
+  time_points <- check_T(time_points, estimate_ME)
   ICC <- check_ICC(ICC)
   RI_cor <- check_RIcor(RI_cor)
-  wSigma <- check_wSigma(wSigma)
+  wSigma <- check_within_cor(within_cor)
   Phi <- check_Phi(Phi)
   reliability <- check_reliability(reliability)
   skewness <- check_skewness(skewness)
   kurtosis <- check_kurtosis(kurtosis)
   alpha <- check_alpha(alpha)
-  est_ME <- check_est_ME(est_ME)
+  estimate_ME <- check_estimate_ME(estimate_ME)
   reps <- check_reps(reps)
   bootstrap_reps <- check_reps(bootstrap_reps)
   seed <- check_seed(seed)
   constraints <- check_constraints(constraints)
   bounds <- check_bounds(bounds, constraints)
   estimator <- check_estimator(estimator, skewness, kurtosis)
+  save_path <- check_save_path(save_path)
 
   # Compute population parameter values for data generation
   Psi <- compute_Psi(Phi, wSigma)
@@ -150,7 +156,7 @@ powRICLPM <- function(target_power,
   }
 
   # Check arguments II
-  sample_size <- check_N(sample_size, time_points, constraints, est_ME)
+  sample_size <- check_N(sample_size, time_points, constraints, estimate_ME)
   Psi <- check_Psi(Psi)
 
   # Update
@@ -171,14 +177,15 @@ powRICLPM <- function(target_power,
     reliability = reliability,
     skewness = skewness,
     kurtosis = kurtosis,
-    est_ME = est_ME,
+    estimate_ME = estimate_ME,
     alpha = alpha,
     reps = reps,
     bootstrap_reps = bootstrap_reps,
     seed = seed,
     constraints = constraints,
     bounds = bounds,
-    estimator = estimator
+    estimator = estimator,
+    save_path
   )
 
   # Update
@@ -198,6 +205,7 @@ powRICLPM <- function(target_power,
     reps = reps,
     bootstrap_reps = bootstrap_reps,
     constraints = constraints,
+    save_path = save_path,
     .options = furrr::furrr_options(
       seed = seed,
       scheduling = 2L # Dynamic scheduling
@@ -231,7 +239,6 @@ powRICLPM <- function(target_power,
 #' # Example - Create Mplus syntax to simulate power across range of sample sizes
 #' # Define population parameters for lagged effects and within-component correlations
 #' Phi <- matrix(c(.4, .1, .2, .3), ncol = 2, byrow = TRUE)
-#' wSigma <- matrix(c(1, .3, .3, 1), ncol = 2, byrow = TRUE)
 #'
 #' # Create and save Mplus model syntax
 #' \dontrun{
@@ -243,7 +250,7 @@ powRICLPM <- function(target_power,
 #'   ICC = c(0.4, 0.5, 0.6),
 #'   RI_cor = 0.3,
 #'   Phi = Phi,
-#'   wSigma = wSigma,
+#'   within_cor = 0.3,
 #'   reps = 10000,
 #'   seed = 1234
 #' )
@@ -257,7 +264,7 @@ powRICLPM_Mplus <- function(search_lower = NULL,
                             ICC,
                             RI_cor,
                             Phi,
-                            wSigma,
+                            within_cor,
                             reps = 1000,
                             seed = NA,
                             save_path = getwd(),
@@ -266,7 +273,7 @@ powRICLPM_Mplus <- function(search_lower = NULL,
   time_points <- check_T(time_points)
   ICC <- check_ICC(ICC)
   RI_cor <- check_RIcor(RI_cor)
-  wSigma <- check_wSigma(wSigma)
+  wSigma <- check_within_cor(within_cor)
   Phi <- check_Phi(Phi)
   reps <- check_reps(reps)
   seed <- check_seed(seed)
@@ -305,7 +312,8 @@ powRICLPM_Mplus <- function(search_lower = NULL,
 
   # Inform user
   message(rlang::format_error_bullets(c(
-    i = "Mplus model syntax created:")))
+    i = "Mplus model syntax created:"
+  )))
   cat("\n\n  Directory:", save_path)
   cat("\n  Sample size(s):", sample_size)
   cat("\n  Number of time points:", time_points)