+++ tests/testthat/test-Mplus.R
@@ -2,7 +2,6 @@ test_that("Mplus model syntax creation works", {
 
   # Create valid powRICLPM_Mplus() input
   Phi <- matrix(c(.5, .1, .4, .5), ncol = 2, byrow = TRUE)
-  wSigma <- matrix(c(1, .3, .3, 1), ncol = 2, byrow = TRUE)
   save_path <- getwd()
 
   # Create Mplus input file for RI-CLPM MCMC power analysis
@@ -12,7 +11,7 @@ test_that("Mplus model syntax creation works", {
     ICC = 0.5,
     RI_cor = 0.3,
     Phi = Phi,
-    wSigma = wSigma,
+    within_cor = 0.3,
     reps = 10000,
     seed = 123456
   )