+++ tests/testthat/test-lavaan.R
@@ -14,20 +14,20 @@ test_that("lavaan model syntax creation works", {
   wSigma <- matrix(c(1, .3, .3, 1), ncol = 2, byrow = TRUE)
   Psi <- matrix(c(0.71, -0.037, -0.037, 0.47), ncol = 2, byrow = TRUE)
   constraints <- "none"
-  est_ME <- FALSE
+  estimate_ME <- FALSE
   skewness <- 0
   kurtosis <- 0
   alpha <- 0.05
 
   # Generate lavaan parameter table
-  lav_out <- powRICLPM:::create_lavaan(cond, Phi, wSigma, Psi, constraints, est_ME, skewness, kurtosis, alpha)
+  lav_out <- powRICLPM:::create_lavaan(cond, Phi, wSigma, Psi, constraints, estimate_ME, skewness, kurtosis, alpha)
 
   # Test
   expect_type(lav_out, "list")
   expect_equal(length(lav_out), 18)
-  expect_equal(names(lav_out), c("sample_size", "time_points", "ICC", "RI_var", "RI_cov", "pop_synt", "pop_tab",
-                                 "est_synt", "est_tab", "est_ME", "skewness", "kurtosis", "alpha", "estimates", "uncertainty",
-                                 "errors", "not_converged", "inadmissible"))
-
-
+  expect_equal(names(lav_out), c(
+    "sample_size", "time_points", "ICC", "RI_var", "RI_cov", "pop_synt", "pop_tab",
+    "est_synt", "est_tab", "estimate_ME", "skewness", "kurtosis", "alpha", "estimates", "uncertainty",
+    "errors", "not_converged", "inadmissible"
+  ))
 })