+++ tests/testthat/test-summary.R
@@ -2,7 +2,6 @@ test_that("powRICLPM_summary() works", {
 
   # Create valid powRICLPM() input
   Phi <- matrix(c(0.4, 0.15, 0.2, 0.3), ncol = 2, byrow = TRUE)
-  wSigma <- matrix(c(1, 0.3, 0.3, 1), ncol = 2, byrow = TRUE)
 
   # Base condition - Single experimental condition
   out1 <- powRICLPM(
@@ -12,7 +11,7 @@ test_that("powRICLPM_summary() works", {
     ICC = 0.5,
     RI_cor = 0.3,
     Phi = Phi,
-    wSigma = wSigma,
+    within_cor = 0.3,
     reps = 2,
     seed = 123456
   )