+++ NEWS.md
@@ -1,3 +1,18 @@
+This file tracks changes to the package.
+
+# powRICLPM 0.1.0
+
+## New features 
+
+* `powRICLPM` can now save the generated data sets by specifying a path with the `save_dat = ""` argument
+
+## Minor improvements and fixes
+
+* The `est_ME` argument in `powRICLPM` has been renamed `estimate_ME`. 
+* Internal fitting model fitting using `lavaan` now skips certain checks to speed up the process. 
+* The `wSigma` argument in `powRICLPM` has been replaced with the `within_cor` argument. Now, only a `double` denoting the correlation between the within-components needs to specified rather than a correlation matrix. 
+* By default, `powRICLPM` now discards results from Monte Carlo replications with inadmissible parameter results, unless bounded estimation is used (`bounds = TRUE`). 
+
 # powRICLPM 0.0.0.9004
 
 ## New features